LayoutTests/accessibility/ax-thread-text-apis/display-contents-end-text-marker.html

<!DOCTYPE HTML><!-- webkit-test-runner [ runSingly=true AccessibilityThreadTextApisEnabled=true ] -->
<!-- Remove this in favor of display-contents/end-text-marker.html after AX_THREAD_TEXT_APIS enabled is the default. -->
<html>
<head>
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<style>
td, th { display: contents; }
</style>
</head>
<body id="body" role="group">

<div id="test-contents">
    Foo text

    <table>
    <caption>This is a table caption</caption>
    <thead>
        <tr>
            <th>Author</th>
            <th>Title</th>
            <th>Year</th>
        </tr>
    </thead>
    <tbody>
        <tr>
            <td>Stephen Hawking</td>
            <td>A Brief History of Time</td>
            <td>1988</td>
        </tr>
        <tr>
            <td>Carl Sagan</td>
            <td>Cosmos</td>
            <td>1980</td>
        </tr>
        <tr>
            <td>Will Gater</td>
            <td>The Mysteries of the Universe</td>
            <td>2020</td>
        </tr>
    </tbody>
    </table>
</div>

<script>
var output = "This test ensures that computing text markers includes all display:contents text content.\n\n";

if (window.accessibilityController) {
    const webArea = accessibilityController.rootElement.childAtIndex(0);

    const startMarker = webArea.startTextMarker;
    const endMarker = webArea.endTextMarker;
    const textMarkerRange = webArea.textMarkerRangeForMarkers(startMarker, endMarker);

    // Fails because:
    //  1. We fail to emit a newline before entering a table, right after "Foo text" (TextIterator::advance() -> TextIterator::handleNonTextNode() -> TextIterator::representNodeOffsetZero() -> shouldEmitNewlineBeforeNode() returns true)
    //  2. We don't emit spaces between cells
    //  3. We miss the table caption text entirely (it is rendered and selectable text, so we should be including it)
    output += `Web area string for start-to-end text marker range:\n\n${webArea.stringForTextMarkerRange(textMarkerRange)}\n`;
    document.getElementById("test-contents").style.display = "none";
    debug(output);
}
</script>
</body>
</html>
